Example of simulated incidence table
For this particular example, we used the Gamma distribution to generate the mean abundances of the species in the community and Poisson distributions with the given means to generate abundance data. Then, abundances were converted to incidence data.
This file contains 100 replicates. Each replicate consist of an incidence table with 100 species plus 10 error species (each error species has a mean abundance given by the quantile of probability 0.000001).